pyproject.toml
[tool.poetry]
name = "slack-openrouter-agent"
version = "0.1.0"
description = "A Slack bot that uses OpenRouter to answer questions"
packages = [
    { include = "flask_app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
slack-sdk = "^3.27.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
aiohttp = "^3.11.18"
asyncio = "^3.4.3"
flask = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
